+++ src/paths/paths.jl
@@ -814,7 +814,9 @@ include("segments/bspline_optimization.jl")
 
 include("routes.jl")
 
+include("channel_routing_core.jl")
 include("channels.jl")
+include("channel_autorouter.jl")
 
 function change_handedness!(seg::Union{Turn, Corner})
     return seg.α = -seg.α